Back-End W/Arrays Added. UI Blueprints Added
Finished setting-up the entire back-end, including the text file as well as access to the text file and access to the array holders that will allow searching and sorting through them. Implemented sorting as well as linear/binary searches and two classes that will hold that info in order to output it to the proper text file. Also set-up actually outputting to that text-file. In summary, what's left to implement is the UI, Linked-List variations of all of the searches and sorts, Hashing for array holders, and creating some text files that will work with String objects instead of integers
(though the current implementation SHOULD work with data types other than ints, though they will still write into the wrong text files).